+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.6.12
 
 * Base AC effects no longer multiply if retain effects is selected.
+* Support for Life Domain bonus with and without Active Effects.
 
 # 0.6.11